APB-1865 Move Invitation Json.Format for mongo to Repository file
