+++ app/uk/gov/hmrc/agentclientauthorisation/model/Invitation.scala
@@ -164,57 +164,6 @@ object Invitation {
 
 }
 
-object InvitationMongoFormat {
-
-  import play.api.libs.json.{ JsPath, Reads }
-  import play.api.libs.functional.syntax._
-
-
-  implicit val serviceFormat = Service.format
-  implicit val statusChangeEventFormat = Json.format[StatusChangeEvent]
-  implicit val oidFormats = ReactiveMongoFormats.objectIdFormats
-
-  def read(id: BSONObjectID, invitationId: InvitationId, arn: Arn, service: Service, clientId: String, clientIdType: String,
-            suppliedClientId: String, suppliedClientIdType: String, postcode: Option[String], expiryDateOp: Option[LocalDate],
-            events: List[StatusChangeEvent]): Invitation = {
-    val expiryDate = expiryDateOp.getOrElse(events.head.time.plusDays(10).toLocalDate)
-
-    Invitation(id, invitationId, arn, service, ClientIdentifier(clientId, clientIdType),
-      ClientIdentifier(suppliedClientId, suppliedClientIdType), postcode, expiryDate, events)
-  }
-
-  val reads: Reads[Invitation] = (
-      (JsPath \ "id").read[BSONObjectID] and
-      (JsPath \ "invitationId").read[InvitationId] and
-      (JsPath \ "arn").read[Arn] and
-      (JsPath \ "service").read[Service] and
-      (JsPath \ "clientId").read[String] and
-      (JsPath \ "clientIdType").read[String] and
-      (JsPath \ "suppliedClientId").read[String] and
-      (JsPath \ "suppliedClientIdType").read[String] and
-      (JsPath \ "postcode").readNullable[String] and
-      (JsPath \ "expiryDate").readNullable[LocalDate] and
-      (JsPath \ "events").read[List[StatusChangeEvent]]) (read _)
-
-  val writes  = new Writes[Invitation] {
-    def writes(invitation: Invitation) = Json.obj(
-      "id" -> invitation.id,
-      "invitationId" -> invitation.invitationId,
-      "arn" -> invitation.arn.value,
-      "service" -> invitation.service.id,
-      "clientId" -> invitation.clientId.value,
-      "clientIdType" -> invitation.clientId.typeId,
-      "postcode" -> invitation.postcode,
-      "suppliedClientId" -> invitation.suppliedClientId.value,
-      "suppliedClientIdType" -> invitation.suppliedClientId.typeId,
-      "events" -> invitation.events,
-      "expiryDate" -> invitation.expiryDate
-    )
-  }
-
-  val mongoFormats = ReactiveMongoFormats.mongoEntity(Format(reads, writes))
-}
-
 /** Information provided by the agent to offer representation to HMRC */
 case class AgentInvitation(
                             service: String,
